tools: liveinstaller: refactor as a lib instead of main package
