+++ toolkit/tools/pkg/image/liveinstaller/config.go
@@ -1,0 +1,11 @@
+package liveinstaller
+
+type Config struct {
+	ConfigFile         string
+	TemplateConfigFile string
+	ForceAttended      bool
+	ImagerTool         string
+	BuildDir           string
+	BaseDirPath        string
+	ImagerLogFile      string
+}
