+++ toolkit/tools/pkg/imagegen/attendedinstaller/primitives/navigationbar/navigationbar.go
@@ -7,7 +7,7 @@ import (
 	"github.com/gdamore/tcell"
 	"github.com/rivo/tview"
 
-	"github.com/microsoft/CBL-Mariner/toolkit/tools/imagegen/attendedinstaller/uitext"
+	"github.com/microsoft/CBL-Mariner/toolkit/tools/pkg/imagegen/attendedinstaller/uitext"
 )
 
 const (
